+++ CHANGELOG.md
@@ -5,6 +5,41 @@
 
 > All notable changes to this project will be documented in this file
 
+## [2.7.0-beta.3](https://github.com/open-sauced/app/compare/v2.7.0-beta.2...v2.7.0-beta.3) (2024-03-12)
+
+
+### 🍕 Features
+
+* transfer contributor insights to different workspaces ([#2917](https://github.com/open-sauced/app/issues/2917)) ([9c51406](https://github.com/open-sauced/app/commit/9c5140653b6042299d6df168400d810cc301677e))
+
+## [2.7.0-beta.2](https://github.com/open-sauced/app/compare/v2.7.0-beta.1...v2.7.0-beta.2) (2024-03-12)
+
+
+### 🐛 Bug Fixes
+
+* transfer hitting undefined ([#2922](https://github.com/open-sauced/app/issues/2922)) ([4db168b](https://github.com/open-sauced/app/commit/4db168b8984ae89c055336b7e6d3f7434bd80edf))
+
+## [2.7.0-beta.1](https://github.com/open-sauced/app/compare/v2.6.1-beta.2...v2.7.0-beta.1) (2024-03-12)
+
+
+### 🍕 Features
+
+* transfer repository insights to other workspace ([#2912](https://github.com/open-sauced/app/issues/2912)) ([23302e3](https://github.com/open-sauced/app/commit/23302e30837223e86f02542c1f2b5a790fd4b62e))
+
+## [2.6.1-beta.2](https://github.com/open-sauced/app/compare/v2.6.1-beta.1...v2.6.1-beta.2) (2024-03-12)
+
+
+### 🐛 Bug Fixes
+
+* redirect to workspaces dashboard after settings update ([#2904](https://github.com/open-sauced/app/issues/2904)) ([9593d17](https://github.com/open-sauced/app/commit/9593d17b7e71e4b690b87e57b8b7cc7d8433f49d))
+
+## [2.6.1-beta.1](https://github.com/open-sauced/app/compare/v2.6.0...v2.6.1-beta.1) (2024-03-11)
+
+
+### 🐛 Bug Fixes
+
+* metrics cards now span full width (and look nicer) ([#2914](https://github.com/open-sauced/app/issues/2914)) ([4ac448d](https://github.com/open-sauced/app/commit/4ac448d30088e2522b86b5c22d2149860ef6dcde))
+
 ## [2.6.0](https://github.com/open-sauced/app/compare/v2.5.1...v2.6.0) (2024-03-11)